feat(common): Message新增枚举类Channel（通道类型）
